+++ packages/client/src/directives/tooltip.ts
@@ -7,10 +7,11 @@ import { popup, alert } from '@/os';
 
 const start = isTouchUsing ? 'touchstart' : 'mouseover';
 const end = isTouchUsing ? 'touchend' : 'mouseleave';
-const delay = 100;
 
 export default {
 	mounted(el: HTMLElement, binding, vn) {
+		const delay = binding.modifiers.noDelay ? 0 : 100;
+
 		const self = (el as any)._tooltipDirective_ = {} as any;
 
 		self.text = binding.value as string;